# docs/swagger/gen/user/v1/swagger.yaml
components:
    schemas:
...
        request.ForgotAuthPasswordRequest:
            properties:
                email:
                    description: メールアドレス
                    type: string
            type: object
...
        request.RefreshAuthTokenRequest:
            properties:
                refreshToken:
                    description: 更新トークン
                    type: string
            type: object
        request.ResetAuthPasswordRequest:
            properties:
                email:
                    description: メールアドレス
                    type: string
                password:
                    description: パスワード
                    type: string
                passwordConfirmation:
                    description: パスワード (確認用)
                    type: string
                verifyCode:
                    description: 検証コード
                    type: string
            type: object
        request.SignInRequest:
            properties:
                password:
                    description: パスワード
                    type: string
                username:
                    description: ユーザー名 (メールアドレス,電話番号)
                    type: string
            type: object
...
        request.UpdateAuthPasswordRequest:
            properties:
                newPassword:
                    description: 新しいパスワード
                    type: string
                oldPassword:
                    description: 現在のパスワード
                    type: string
                passwordConfirmation:
                    description: パスワード (確認用)
                    type: string
            type: object
...
        response.AuthGoogleAccountResponse:
            properties:
                url:
                    description: Googleアカウント連携URL
                    type: string
            type: object
        response.AuthLINEAccountResponse:
            properties:
                url:
                    description: LINEアカウント連携URL
                    type: string
            type: object
        response.AuthResponse:
            properties:
                accessToken:
                    description: アクセストークン
                    type: string
                expiresIn:
                    description: 有効期限
                    type: integer
                refreshToken:
                    description: 更新トークン
                    type: string
                tokenType:
                    description: トークン種別
                    type: string
                userId:
                    description: ユーザーID
                    type: string
            type: object
...
        util.ErrorResponse:
            properties:
                detail:
                    description: エラー詳細
                    type: string
                message:
                    description: エラー概要
                    type: string
                status:
                    description: ステータスコード
                    type: integer
            type: object
    securitySchemes:
        bearerauth:
            bearerFormat: JWT
            scheme: bearer
            type: http
        cookieauth:
            in: cookie
            name: session_id
            type: apiKey
...
info:
    description: 購入者向けのふるマルAPIです。
    title: ふるマル API - 購入者向け
    version: ""
openapi: 3.1.0
paths:
...
    /auth:
        delete:
            description: ふるマルからサインアウトします。
            responses:
                "204":
                    description: No Content
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            security:
                - bearerauth: []
            summary: サインアウト
            tags:
                - Auth
        get:
            description: 認証トークンを検証し、認証情報を取得します。
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.AuthResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            security:
                - bearerauth: []
            summary: トークン検証
            tags:
                - Auth
        post:
            description: ユーザー名/メールアドレスとパスワードでサインインします。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.SignInRequest'
                description: サインイン
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.AuthResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            summary: サインイン
            tags:
                - Auth
    /auth/forgot-password:
        post:
            description: パスワード再設定のための検証コードをメールで送信します。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.ForgotAuthPasswordRequest'
                description: パスワード再設定リクエスト
                required: true
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
            summary: パスワード再設定リクエスト
            tags:
                - Auth
    /auth/forgot-password/verified:
        post:
            description: 検証コードを使用してパスワードを再設定します。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.ResetAuthPasswordRequest'
                description: パスワード再設定実行
                required: true
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
            summary: パスワード再設定実行
            tags:
                - Auth
    /auth/google:
        get:
            description: Google OAuth認証のための認証URLを取得します。
            parameters:
                - description: ステート
                  in: query
                  name: state
                  schema:
                    type: string
                - description: リダイレクトURI
                  in: query
                  name: redirectUri
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.AuthGoogleAccountResponse'
                    description: OK
            security:
                - cookieauth: []
            summary: Google認証URL取得
            tags:
                - Auth
    /auth/line:
        get:
            description: LINE OAuth認証のための認証URLを取得します。
            parameters:
                - description: ステート
                  in: query
                  name: state
                  schema:
                    type: string
                - description: リダイレクトURI
                  in: query
                  name: redirectUri
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.AuthLINEAccountResponse'
                    description: OK
            security:
                - cookieauth: []
            summary: LINE認証URL取得
            tags:
                - Auth
    /auth/password:
        patch:
            description: 現在のパスワードを使用して新しいパスワードに変更します。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.UpdateAuthPasswordRequest'
                description: パスワード変更
                required: true
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            security:
                - bearerauth: []
            summary: パスワード変更
            tags:
                - Auth
    /auth/refresh-token:
        post:
            description: リフレッシュトークンを使用してアクセストークンを更新します。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.RefreshAuthTokenRequest'
                description: トークンリフレッシュ
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.AuthResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            summary: トークンリフレッシュ
            tags:
                - Auth
...
servers:
    - description: 検証環境
      url: https://api.furumaru-stg.and-period.work
    - description: 本番環境
      url: https://api.furumaru.and-period.co.jp